system commands: add core id in the tasklist
FreeRTOS esp additions supports tracking of the core_id by meand of a configuration
option: CONFIG_TASKLIST_INCLUDE_COREID

If this is set, the core id will be printed in an extra column in the
task overview. The following logic is applied:
 0 = core 0
 1 = core 1
 ? = unknown

In case core affinity is not defined at compile time. This happens if
you call xTaskCreate directly in stead of xTaskCreatePinnedToCore.